Extract search functionality to its own component
